separate style.go
